+++ src/diff.js
@@ -6,18 +6,12 @@ import RequiresAsyncError from './requiresAsyncError';
 
 
 function diffElements(actualAdapter, expectedAdapter, actual, expected, expect, options) {
-    let result;
+    
     try {
         return SyncDiff.diffElements(actualAdapter, expectedAdapter, actual, expected, expect, options);
     } catch (e) {
         if (e instanceof RequiresAsyncError) {
-            const promiseResult = AsyncDiff.diffElements(actualAdapter, expectedAdapter, actual, expected, expect, options);
-
-            return promiseResult.then(result => {
-                return result;
-            }).catch(e => {
-                throw e;
-            });
+            return AsyncDiff.diffElements(actualAdapter, expectedAdapter, actual, expected, expect, options);
         }
         throw e;
     }